JIRA table support
